Fix + Ajout fonc + Lha

+++ src/fr/farmeurimmo/verymc/shopgui/GenAmoutShopGui.java
@@ -67,7 +67,7 @@ public class GenAmoutShopGui {
 		}
 		custom4.setItemMeta(customd);
 		
-		int price = 0;
+		Float price = (float) 0;
 		ItemMeta tempameta = aa.getItemMeta();
 		tempameta.setLore(null);
 		if(a == true) {
